==> models/Maxent/scripts/2_variable_selection_and_regularization.Rmd <==
---
title: "Variable Selection and the Regularization Parameter"
author: "Katharine Egan"
email: "katharineegan34@gmail.com"
date: "July 2018"
output: html_document
---

Use the MaxentVariableSelection R package - tests a bunch of beta multipliers and chooses the variables that make up the best model (while simultaneously running a correlation analysis)
https://github.com/alj1983/MaxentVariableSelection


More documentation on using the MaxentVariableSelection package 
https://cran.r-project.org/web/packages/MaxentVariableSelection/vignettes/MaxentVariableSelection.pdf

Libraries: 
```{r}
library(raster)
library(rgeos)
library(rgdal)
library(sp)
library(dismo)
library(tidyverse)
library(readxl)
library(writexl)
library(MaxentVariableSelection)
library(beepr)
```

Maxent models variable selection
```{r}
# Specify the file path to the maxent.jar file, which has to be downloaded from here. Not that you can not just copy the following line but instead have to specify the file path to the folder containing maxent.jar on your own computer.
maxent <- ("C:/maxent/maxent.jar")

# Filepath to the output directory, including the name of the directory, like:
# MAKE SURE YOUR OUTPUT FOLDER IS EMPTY BEFORE RUNNING THIS SCRIPT
outdir <- ("E:/SERVIR/Landslide_Models/models/Maxent/output/2_MaxentVariableSelection_output")

# Here, you specify the filepath to the folder containing all your ASCII grids of environmental variables that you consider to be potentially relevant in setting distribution limits of your target species. All variables must have the same extent and resolution.
gridfolder <- ("E:/SERVIR/Landslide_Models/predictors_ascii") 

# Here, you need to specify the filepath to the csv file of occurrence locations
occurrencelocations <- ("E:/SERVIR/Landslide_Models/models/Maxent/points/occurrences.csv") 

# Specify filepath to the csv of background locations 
backgroundlocations <- ("E:/SERVIR/Landslide_Models/models/Maxent/points/backgroundpoints.csv") 

# Maxent arguments can be specified with the additionalargs argument. You find an overview of possible arguments in the end of the Maxent help file.
# TOGGLE categorical data like this
additionalargs <- c("togglelayertype=land_cover")

# This sets the threshold of model contribution below which environmental variables are excluded from the Maxent model. Model contributions range from 0% to 100% and reflect the importance of environmental variables in limiting the distribution of the target species. In this example, we set the threshold to 5%, which means that all variables will be excluded when they contribute less than 5% to the model:
contributionthreshold <- 5

# This sets the threshold of Pearsons correlation coefficient (ranging from 0 to 1) above which environmental variables are regarded to be correlated (based on values at all background locations). Of the correlated variables, only the variable with the highest contribution score will be kept, all other correlated variables will be excluded from the Maxent model. Correlated variables should be removed because they may reflect the same environmental conditions, and can lead to overly complex or overpredicted models. Also, models compiled with correlated variables might give wrong predictions in scenarios where the correlations between the variables differ.
correlationthreshold <- 0.6

#  This argument sets the values of beta multipliers (regularization multipliers) for which variable selection shall be performed. The smaller this value, the more closely will the projected distribution fit to the training data set.
betamultiplier <- c(1, 3)

# run the variable select function
VariableSelection(maxent, outdir, gridfolder, occurrencelocations,
                  backgroundlocations, additionalargs, 
                  contributionthreshold,
                  correlationthreshold, betamultiplier)

# make a noise when script is done 
beep(3)
```
